pyproject.toml
[tool.poetry]
name = "FastAPI-Auth"
version = "0.1.0"
description = "full authentication system with JWT, OAuth2, 2FA, role-based access control (RBAC), and audit logging for FastAPI applications"
authors = ["Ajiboye Adeleye Pius"]
license = "ISC"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.11"
uvicorn = "^0.34.0"
python-jose = {extras = ["cryptography"], version = "^3.4.0"}
passlib = "^1.7.4"
bcrypt = "^4.3.0"
python-multipart = "^0.0.20"
email-validator = "^2.2.0"
pyotp = "^2.9.0"
SQLAlchemy = "^2.0.38"
databases = "^0.9.0"
uuid7 = "^0.1.0"
pydantic-settings = "^2.8.1"
python-decouple = "^3.8"
alembic = "^1.14.1"
PyMySQL = "^1.1.1"
jwt = "^1.3.1"
PyJWT = "^2.10.1"
cryptography = "^44.0.2"
user-agents = "2.2.0"
Jinja2 = "^3.1.6"
mjml = "^0.11.0"
fastapi-mail = "^1.4.2"
slowapi = "^0.1.9"
premailer = "^3.10.0"
qrcode = "^8.1"
twilio = "^9.5.2"
pillow = "^11.2.1"
geoip2 = "^5.0.1"
APScheduler = "^3.11.0"
Authlib = "^1.5.2"
httpx = "^0.28.1"
itsdangerous = "^2.2.0"
pycountry-convert = "^0.7.2"
pytz = "^2025.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
